projects/PoCs/Kucia_Trochimiak_Grabek/requirements.txt
openai

PyMuPDF
nltk
tqdm

sentence_transformers
chromadb
spacy
cupy-cuda12x
torch

flask
